+++ messages/sendMessage/sendMessage.resolvers.js
@@ -1,0 +1,86 @@
+import e from "express";
+import client from "../../client";
+import { NEW_MESSAGE } from "../../constans";
+import pubsub from "../../pubsub";
+import { protectedResolver } from "../../users/users.utils";
+
+export default {
+  Mutation: {
+    sendMessage: protectedResolver(
+      async (_, { payload, roomId, userId }, { loggedInUser }) => {
+        let room = null;
+        if (userId) {
+          const user = await client.user.findUnique({
+            where: {
+              id: userId,
+            },
+            select: {
+              id: true,
+            },
+          });
+
+          if (!user) {
+            return {
+              ok: false,
+              error: "This user does not exists.",
+            };
+          }
+
+          room = await client.room.create({
+            data: {
+              users: {
+                connect: [
+                  {
+                    id: userId,
+                  },
+                  {
+                    id: loggedInUser.id,
+                  },
+                ],
+              },
+            },
+          });
+        } else if (roomId) {
+          room = await client.room.findUnique({
+            where: {
+              id: roomId,
+            },
+            select: {
+              id: true,
+            },
+          });
+
+          if (!room) {
+            return {
+              ok: false,
+              error: "Room not found.",
+            };
+          }
+        }
+        // 신규메세지 생성
+        const message = await client.message.create({
+          data: {
+            payload,
+            read: false,
+            room: {
+              connect: {
+                id: room.id,
+              },
+            },
+            user: {
+              connect: {
+                id: loggedInUser.id,
+              },
+            },
+          },
+        });
+
+        console.log(message);
+        pubsub.publish(NEW_MESSAGE, { roomUpdates: message });
+        return {
+          ok: true,
+        };
+      }
+    ),
+  },
+};